% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat2sccnvseg.R
\name{ascat2sccnvseg}
\alias{ascat2sccnvseg}
\title{ascat2sccnvseg}
\usage{
ascat2sccnvseg(ascat_summary)
}
\arguments{
\item{ascat_summary}{a segment matrix loaded in from an ASCAT.summary.csv file}
}
\value{
a segment matrix in scCNV format
}
\description{
Converts ASCAT summary segmentation format to scCNV format, for plotting ASCAT data with G_T_chr_plots
}
