% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_correlate_CN_expression.R
\name{bulk_correlate_CN_expression}
\alias{bulk_correlate_CN_expression}
\title{bulk_correlate_CN_expression}
\usage{
bulk_correlate_CN_expression(
  normalised_counts_matrix,
  dna_segments_dir,
  output_dir
)
}
\arguments{
\item{normalised_counts_matrix}{a normalised counts matrix. Must contain metadata columns
like featurecounts matrix.}

\item{dna_segments_dir}{a folder containing cell-specific segmentation files produced by scCNV (or from
ASCAT.summary.csv if converted with the ascat2sccnvsegfunction).}

\item{output_dir}{output directory to store tables correlating expression to copy number for each
gene}
}
\value{
Nothing
}
\description{
Writes tables of RNA expression and DNA copy number for each gene, from each cell.
Used to generate the table files needed for the plot_CN_expression() function
}
