% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_G_T_chr_plots.R
\name{bulk_G_T_chr_plots}
\alias{bulk_G_T_chr_plots}
\title{bulk_G_T_chr_plots}
\usage{
bulk_G_T_chr_plots(
  dna_segments_dir = NA,
  rna_segments_dir = NA,
  output_dir = "./",
  chromosomes = NA,
  threads
)
}
\arguments{
\item{dna_segments_dir}{folder of CNV segment files from scCNV or converted from ASCAT with ascat2sccnvseg()}

\item{rna_segments_dir}{folder of expression segment files, from get_expression_by_segment()}

\item{output_dir}{directory where the plots will be saved}

\item{threads}{number of threads to run concurrently (defaults to number of avaliable cores minus 1)}
}
\value{
Nothing
}
\description{
runs the G_T_chr_plot function for a whole directory of scCNV segment files and RNA segment
}
