% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_get_expression_by_segment.R
\name{bulk_get_expression_by_segment}
\alias{bulk_get_expression_by_segment}
\title{barman}
\usage{
bulk_get_expression_by_segment(counts_matrix, dna_segments_dir, output_dir)
}
\arguments{
\item{counts_matrix}{a featureCounts matrix}

\item{dna_segments_dir}{folder of CNV segment files from scCNV or converted from ASCAT with ascat2sccnvseg()}

\item{output_dir}{an output directory to store RNAsegment files}
}
\value{
Nothing
}
\description{
runs the get_expression_by_segment function for a whole directory of scCNV segment files
}
