% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_boxplots.R
\name{expression_boxplots}
\alias{expression_boxplots}
\title{barman}
\usage{
expression_boxplots(experimental_group, control_group, counts_matrix)
}
\arguments{
\item{experimental_group}{list of column names corresponding to samples we want to use to show the mean expression of cells of interest}

\item{control_group}{list of column names corresponding to samples we want to use to show the mean expression over control cells}

\item{counts_matrix}{counts matrix with expression we want to plot, needs to already be normalised}
}
\value{
Nothing
}
\description{
Generates per chromosome boxplots for the mean expression for each gene, grouped by chromosome
}
