% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_normalise_scrna.R
\name{filter_and_normalise_scRNA}
\alias{filter_and_normalise_scRNA}
\title{barman}
\usage{
filter_and_normalise_scRNA(counts_matrix, annotation_table)
}
\arguments{
\item{counts_matrix}{a raw counts matrix from featureCounts}

\item{annotation_table}{an annotation table}
}
\value{
Nothing
}
\description{
filters out poor quality cells and features, and converts counts matrix to FPKM
}
