% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_normalise_scrna__seurat.R
\name{filter_and_normalise_scRNA__seurat}
\alias{filter_and_normalise_scRNA__seurat}
\title{filter_and_normalise_scRNA__seurat}
\usage{
filter_and_normalise_scRNA__seurat(
  counts_matrix,
  output_dir = "./",
  annotation_table = FALSE,
  return_sce = FALSE,
  manual_filter = FALSE,
  filter_only = FALSE
)
}
\arguments{
\item{counts_matrix}{a raw counts matrix from featureCounts}

\item{output_dir}{output directory to save plots}

\item{annotation_table}{an annotation table}

\item{return_sce}{boolean to return SCE object instead of matrix}

\item{manual_filter}{vector of 4 values, manually specifying upper cutoffs to apply for total counts. e.g. c(20000,6000000,20,50) for upper limits of total counts, total_features_by_counts, pct mt, and pct ercc respectively. Also accepts lists instead of ints, where the first element of list is lower cutoff and second is upper cutoff.}

\item{filter_only}{boolean to return only filtered (i.e. non-fpkm normalized) matrix}
}
\value{
seurat normalized counts matrix (returns 'logcounts(sce)')
}
\description{
filters out poor quality cells and features, and converts counts matrix to seurat normalized matrix
}
