% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_SCE_to_Seurat.R
\name{filtered_SCE_to_Seurat}
\alias{filtered_SCE_to_Seurat}
\title{barman}
\usage{
filtered_SCE_to_Seurat(SCE_obj, counts_matrix, annotation_table)
}
\arguments{
\item{SCE_obj}{a SingleCellExperiment object created by running filter_and_normalise_scRNA() with the return_sce option set to TRUE}

\item{counts_matrix}{a raw counts matrix from featureCounts}

\item{annotation_table}{an annotation table with at least one of the columns containing the colnames of the counts matrix}
}
\value{
a normalized Seurat object
}
\description{
Convert FPKM normalized object SCE to Seurat Object
}
