% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_T_chr_plot.R
\name{G_T_chr_plot}
\alias{G_T_chr_plot}
\title{barman}
\usage{
G_T_chr_plot(cnv_data, exp_data, title, save)
}
\arguments{
\item{cnv_data}{a segmentation file produced by scCNV. Can also use an ASCAT.summary.csv if converted with the ascat2sccnvseg() function.}

\item{exp_data}{a segmentation file generated by calling the get_expression_by_segment() function on a featurecounts matrix and on a DNA segmentation file}

\item{title}{Plot title}

\item{save}{boolean to specify if the plot should be generated or saved as PNG. If set to TRUE, the plot will be saved as the specified title.}
}
\value{
a ggplot2 object or nothing if save==TRUE
}
\description{
Generates Genome CNV and Transcriptome Expression plots by chromsome from previously generated segmentation files.
}
