% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_copynumber_for_gene.R
\name{get_copynumber_for_gene}
\alias{get_copynumber_for_gene}
\title{get_copynumber_for_gene}
\usage{
get_copynumber_for_gene(dna_segments, chr, start, end, gene_name)
}
\value{
a list of values corresponding to one segment along with the PCF of the segments
}
\description{
Groups all segments that overlap each gene boundry and returns that to the
G_T_chr_plot which calls this internal function
}
