% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_by_segment.R
\name{get_expression_by_segment}
\alias{get_expression_by_segment}
\title{get_expression_by_segment}
\usage{
get_expression_by_segment(cell_id, counts_matrix, segment_file)
}
\arguments{
\item{cell_id}{the column name in counts_matrix of the cell we want to get avg expression for each segment}

\item{counts_matrix}{a featureCounts matrix}

\item{segment_file}{the CNV segment file from scCNV or converted from ASCAT with ascat2sccnvseg()}
}
\value{
a segmentation matrix of gene expression
}
\description{
create a segment table of RNA expression, using the same segments as a provided CNV segment file
}
