% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_by_ref_group.R
\name{logR_by_ref_group}
\alias{logR_by_ref_group}
\title{logR_by_ref_group}
\usage{
logR_by_ref_group(raw_counts_matrix, reference_cells, manual_filter = FALSE)
}
\arguments{
\item{raw_counts_matrix}{a counts matrix with no normalisation}

\item{reference_cells}{a list of strings corresponding to column names in counts matrix. These are
the names of cells we will use as a reference group}

\item{manual_filter}{optional list of values to pass to the internal filter_and_normalise_scRNA
function to filter out poor quality samples before calculating log fold change}
}
\value{
a counts matrix containing log fold change values between specified reference cells and
all others.
}
\description{
converts raw counts matrix to log fold change values, based on fold change
between mean of specified experiemental group and specified reference group.
}
