% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_chromosome_list.R
\name{order_chromosome_list}
\alias{order_chromosome_list}
\title{order_chromosome_list}
\usage{
order_chromosome_list(random_order_chr_list)
}
\arguments{
\item{random_order_chr_list}{a vector of unordered values either strings or integers}
}
\value{
a vector of sorted values first by numeric order then by alphabetical order
}
\description{
sorts a vector of strings/ints numerically then alphabetically to show chromosmes in
proper order
}
