% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CN_expression.R
\name{plot_CN_expression}
\alias{plot_CN_expression}
\title{plot_CN_expression}
\usage{
plot_CN_expression(
  CN_expression_dir,
  annotation_table = NA,
  annotation_column = NA
)
}
\arguments{
\item{CN_expression_dir}{directory of tables correlating expression to copy number for each}

\item{annotation_table}{an annotation table where rownames are the same cell ids as the prefix in filenames in
CN_expression_dir, and where at least one column of matrix determines cell group / type}

\item{annotation_column}{name of column in annotation_table corresponding to cell group / type}
}
\value{
a ggplot object of boxplot of average expression per copy number (optionally per cell
group)
}
\description{
Plots a boxplot of average expression for each DNA copy number for each gene between
specified groups of cells
}
