% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rna_clusters.R
\name{plot_rna_clusters}
\alias{plot_rna_clusters}
\title{plot_rna_clusters}
\usage{
plot_rna_clusters(
  normalised_counts_matrix,
  annotation_table,
  umap_vars = FALSE
)
}
\arguments{
\item{normalised_counts_matrix}{a filtered, normalized, and log transformed counts matrix}

\item{annotation_table}{a table of cell annotations, with  rows corresponding to colnames in normalised_counts_matrix}

\item{umap_vars}{a vector of 2 colnames from annotation table, to manually specifiy the plot's colour and size values respectively}
}
\value{
ggplot2 object of UMAP
}
\description{
plots UMAP of a filtered and normalized SingleCellExperiement object
}
