% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qc_plots}
\alias{qc_plots}
\title{barman}
\usage{
qc_plots(counts_matrix)
}
\arguments{
\item{counts_matrix}{a raw counts matrix from featureCounts}
}
\value{
ggplot2 object of violin plot
}
\description{
plots total_counts total_features_by_counts, mt% and ercc% for given counts matrix
}
