% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_non_specific_columns.R
\name{remove_non_specific_annotation_columns}
\alias{remove_non_specific_annotation_columns}
\title{barman}
\usage{
remove_non_specific_annotation_columns(annotation_table, exclude)
}
\arguments{
\item{annotation_table}{a table of annotations we want to prune}

\item{exclude}{a list of column names we want to keep even if they are non-specific}
}
\value{
a pruned annotation table
}
\description{
removes columns from table where all values are the same or where all values are distinct strings
}
