% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_counts_matrix_from_annotation.R
\name{rename_counts_matrix_from_annotation}
\alias{rename_counts_matrix_from_annotation}
\title{rename_counts_matrix_from_annotation}
\usage{
rename_counts_matrix_from_annotation(
  counts_matrix,
  annotation_table,
  dna_cell_id,
  rna_cell_id = ""
)
}
\arguments{
\item{counts_matrix}{the featurecounts matrix}

\item{annotation_table}{annotation table containing at least one column with the featurecounts column names, and another column we want to use to rename the counts matrix.}

\item{dna_cell_id}{the columnname in the annotation_table, we want to use as the new featurecounts column names}

\item{rna_cell_id}{(optional) the columnname in the annotation_table containing the current featurecounts columnnames}
}
\value{
a renamed counts matrix
}
\description{
Renames the counts matrix column names from a provided annotation table
}
