% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_segment_files_from_annotation.R
\name{rename_segment_files_from_annotation}
\alias{rename_segment_files_from_annotation}
\title{rename_segment_files_from_annotation}
\usage{
rename_segment_files_from_annotation(
  segment_files_dir,
  annotation_table,
  dna_cell_id
)
}
\arguments{
\item{segment_files_dir}{the folder of files to rename}

\item{annotation_table}{annotation table containing a column of unique cell IDs that can be used to differentiate each segment file}

\item{dna_cell_id}{the columnname in the annotation_table of the column containg the unique cell IDs}
}
\value{
Nothing
}
\description{
Renames a folder of segmentation files to remove the bloat from filenames, leaving only the cell id and a ".txt" suffix
}
