% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logR_by_ref_group.R
\name{logR_by_ref_group}
\alias{logR_by_ref_group}
\title{logR_by_ref_group}
\usage{
logR_by_ref_group(fpkm_counts, reference_cells)
}
\arguments{
\item{fpkm_counts}{a normalized counts matrix}

\item{reference_cells}{a list of strings corresponding to column names in counts matrix. These are
the names of cells we will use as a reference group}
}
\value{
a counts matrix containing log fold change values between specified reference cells and
all others.
}
\description{
converts raw counts matrix to log fold change values, based on fold change
between mean of specified experiemental group and specified reference group.
}
