% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rna_clusters.R
\name{plot_rna_clusters}
\alias{plot_rna_clusters}
\title{barman}
\usage{
plot_rna_clusters(
  normalised_SCE_object,
  regress_out = FALSE,
  umap_vars = FALSE
)
}
\arguments{
\item{normalised_SCE_object}{a filtered and normalized SingleCellExperiement, like the type returned from running filter_and_normalise_scRNA() with the return_sce=T parameter.}

\item{regress_out}{a vector of colnames from annotation table, to regress out of dataset before plotting}

\item{umap_vars}{a vector of colnames from annotation table, to manually specifiy colour_by and size_by values respectively}
}
\value{
ggplot2 object of UMAP
}
\description{
plots UMAP of a filtered and normalized SingleCellExperiement object
}
