% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_transform.R
\name{zscore_transform}
\alias{zscore_transform}
\title{barman}
\usage{
zscore_transform(all_cells_to_normalise, counts_matrix)
}
\arguments{
\item{all_cells_to_normalise}{list of column names that we want to zscore transform}

\item{counts_matrix}{counts matrix with normally distrubted data}
}
\value{
zscore transformed counts matrix
}
\description{
apply a zscore transformation to a list of columns based on average and SD of a column names provided by all_cells_to_normalise arg. Zscore should only be applied on counts matrix containing values that follow a normal distribution.
}
